chore: refactor request configurable since only index use it
